added  prefix to all the api routes imported from the api folder
